requirements.txt
streamlit
sentence-transformers
faiss-cpu
pypdf
python-docx
numpy